Add files via upload
Abstract:

The demand for electricity is increasing day by day, which can’t be fulfilled 
by non-renewable energy sources alone. It has become imperative for the power and 
energy engineers to look out for the renewable energy sources such as sun, wind, 
geothermal, ocean and biomass as sustainable, cost-effective and environment 
friendly alternatives for conventional energy sources. However, the non-availability 
of these renewable energy resources all the time throughout the year has led to 
research in the area of hybrid renewable energy systems. 

The renewable energy sources are emerging options to fulfill the energy 
demand, but unreliable due to the stochastic nature of their occurrence. Hybrid 
renewable energy system (HRES) combines two or more renewable energy sources 
like wind turbines and solar systems. In the past few years, a lot of research has taken 
place in the design, optimization, operation and control of the renewable hybrid 
energy systems. It is indeed evident that this area is still emerging and vast in scope.

The objective of this project is to present a comprehensive review of various 
aspects of HRES. This project discusses pre-feasibility analysis, optimum sizing,
modeling and other factors. Along with this, techno-economic analysis of Hybrid 
Energy Renewable System is done to serve a remote community. The optimization 
of the key performance indicators of the proposed hybrid power system is done. We 
also focused on a feasibility study to integrate a battery energy storage system and 
converter with a hybrid wind-solar standalone system. The results obtained from this 
work are analyzed to select the best options among the available configurations 
based on the lowest TPV and LCOE produced by each configuration. The results 
from several case studies show that incorporation of PV and WTG have reduced the 
operating cost of the system.